apps/face-compare/requirements.txt
fastapi
uvicorn[standard]
deepface
tf-keras
opencv-python-headless
numpy
pycryptodome
requests
av
python-dotenv
